feat(5-2): add PlacesContainer to the project
